remote/remote/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_pytorch_select=0.1=cpu_0
altair=4.1.0=pypi_0
astor=0.8.1=pypi_0
atomicwrites=1.4.0=pyh9f0ad1d_0
attrs=19.3.0=py_0
backcall=0.2.0=pyh9f0ad1d_0
base58=2.0.1=pypi_0
blas=1.0=mkl
bleach=3.1.5=pyh9f0ad1d_0
blinker=1.4=pypi_0
bokeh=2.1.1=py38_0
boto3=1.14.20=pypi_0
botocore=1.17.20=pypi_0
brotlipy=0.7.0=py38h1e8a9f7_1000
bullet=2.88=py38he350917_1001
ca-certificates=2020.6.24=0
cachetools=4.1.1=pypi_0
certifi=2020.6.20=py38_0
cffi=1.14.0=py38h7a1dbc1_0
chardet=3.0.4=py38h32f6830_1006
click=7.1.2=py_0
colorama=0.4.3=py_0
cpuonly=1.0=0
cryptography=2.9.2=py38hba49e27_0
cudatoolkit=10.1.243=h74a9793_0
cxx-compiler=1.0.4=he980bc4_0
cycler=0.10.0=py_2
decorator=4.4.2=py_0
defusedxml=0.6.0=py_0
docutils=0.15.2=pypi_0
entrypoints=0.3=py38h32f6830_1001
enum-compat=0.0.3=pypi_0
et_xmlfile=1.0.1=py38_0
flake8=3.8.3=pyh9f0ad1d_0
flake8-polyfill=1.0.2=py_0
flake8-quotes=3.2.0=pyh9f0ad1d_0
flask=1.1.2=py_0
freetype=2.10.2=hd328e21_0
git=2.23.0=h6bb4b03_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=2.9=py_1
importlib-metadata=1.6.1=py38h32f6830_0
importlib_metadata=1.6.1=0
imutils=0.5.3=py38_1
intel-openmp=2020.1=216
ipykernel=5.3.0=py38h5ca1d4c_0
ipython=7.15.0=py38h32f6830_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.0a0=pypi_0
itsdangerous=1.1.0=py_0
jdcal=1.4.1=py_0
jedi=0.17.1=py38h32f6830_0
jinja2=2.11.2=py_0
jmespath=0.10.0=pypi_0
jpeg=9b=vc14h4d7706e_1
json5=0.9.4=pyh9f0ad1d_0
jsonschema=3.2.0=py38h32f6830_1
jupyter_client=6.1.3=py_0
jupyter_core=4.6.3=py38h32f6830_1
jupyterlab=2.1.5=py_0
jupyterlab_server=1.1.5=py_0
kiwisolver=1.2.0=py38heaebd3c_0
libblas=3.8.0=14_mkl
libcblas=3.8.0=14_mkl
libiconv=1.15=vc14h29686d3_5
liblapack=3.8.0=14_mkl
libopencv=4.0.1=hbb9e17c_0
libpng=1.6.37=h2a8f88b_0
libsodium=1.0.17=h2fa13f4_0
libtiff=4.1.0=h56a325e_1
libxml2=2.9.10=h464c3ec_1
libxslt=1.1.34=he774522_0
lxml=4.5.1=py38h1350720_0
lz4-c=1.9.2=h62dcd97_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markupsafe=1.1.1=py38he774522_0
matplotlib=3.2.1=0
matplotlib-base=3.2.1=py38h1626042_0
mccabe=0.6.1=py_1
mistune=0.8.4=py38h9de7a3e_1001
mkl=2019.4=245
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.1.0=py38h45dec08_0
mkl_random=1.1.0=py38hf9181ef_0
more-itertools=8.4.0=py_0
msys2-conda-epoch=20160418=1
nbconvert=5.6.1=py38h32f6830_1
nbformat=5.0.6=py_0
ninja=1.9.0=py38h74a9793_0
notebook=6.0.3=py38h32f6830_0
numpy=1.18.5=py38h6530119_0
numpy-base=1.18.5=py38hc3f5095_0
olefile=0.46=py_0
opencv=4.0.1=py38h2a7c758_0
openpyxl=3.0.3=py_0
openssl=1.1.1g=he774522_0
packaging=20.4=pyh9f0ad1d_0
pandas=1.0.5=py38h47e9c7a_0
pandas-datareader=0.8.1=py_0
pandoc=2.9.2.1=0
pandocfilters=1.4.2=py_1
parso=0.7.0=pyh9f0ad1d_0
pathtools=0.1.2=pypi_0
pep8-naming=0.10.0=pyh9f0ad1d_0
pickleshare=0.7.5=py38h32f6830_1001
pillow=7.1.2=py38hcc1f983_0
pip=20.1.1=py38_1
pluggy=0.13.1=py38h32f6830_2
prometheus_client=0.8.0=pyh9f0ad1d_0
prompt-toolkit=3.0.5=py_1
protobuf=3.12.2=pypi_0
py=1.8.2=pyh9f0ad1d_0
py-opencv=4.0.1=py38he44ac1e_0
pyautogui=0.9.48=py38_0
pycodestyle=2.6.0=pyh9f0ad1d_0
pycparser=2.20=py_0
pydeck=0.4.0b2=pypi_0
pyflakes=2.2.0=pyh9f0ad1d_0
pygments=2.6.1=py_0
pymodbus=2.3.0=pypi_0
pymsgbox=1.0.7=py_0
pyopenssl=19.1.0=py_1
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.9.2=py38ha925a31_4
pyrsistent=0.16.0=py38h9de7a3e_0
pyscreeze=0.1.26=py_0
pyserial=3.4=pypi_0
pysocks=1.7.1=py38h32f6830_1
pytesseract=0.3.4=pyh9f0ad1d_0
pytest=5.4.3=py38h32f6830_0
python=3.8.3=he1778fa_0
python-dateutil=2.8.1=py_0
python_abi=3.8=1_cp38
pytorch=1.5.1=py3.8_cpu_0
pytweening=1.0.3=py_0
pytz=2020.1=py_0
pywin32=227=py38hfa6e2cd_0
pywinpty=0.5.7=py38_0
pyyaml=5.3.1=py38he774522_1
pyzmq=19.0.1=py38h77b9d75_0
qt=5.9.7=vc14h73c81de_0
requests=2.24.0=pyh9f0ad1d_0
s3transfer=0.3.3=pypi_0
scipy=1.3.2=py38h582fac2_0
send2trash=1.5.0=py_0
setuptools=47.3.1=py38_0
sip=4.19.13=py38ha925a31_0
six=1.15.0=py_0
sqlalchemy=1.3.18=py38he774522_0
sqlite=3.32.3=h2a8f88b_0
streamlit=0.62.1=pypi_0
terminado=0.8.3=py38h32f6830_1
testpath=0.4.4=py_0
tk=8.6.10=he774522_0
toml=0.10.1=pypi_0
toolz=0.10.0=pypi_0
torchvision=0.6.1=py38_cpu
tornado=6.0.4=py38he774522_1
traitlets=4.3.3=py38h32f6830_1
typing_extensions=3.7.4.2=py_0
tzlocal=2.0.0=pypi_0
urllib3=1.25.9=py_0
validators=0.15.0=pypi_0
vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_2
vs2015_win-64=14.0.25420=h55c1224_11
watchdog=0.10.2=pypi_0
wcwidth=0.2.5=pyh9f0ad1d_0
webencodings=0.5.1=py_1
werkzeug=1.0.1=py_0
wheel=0.34.2=py38_0
widgetsnbextension=4.0.0a0=pypi_0
win_inet_pton=1.1.0=py38_0
wincertstore=0.2=py38_0
winpty=0.4.3=4
xlrd=1.2.0=py_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=he774522_0
zeromq=4.3.2=h6538335_2
zipp=3.1.0=py_0
zlib=1.2.11=vc14h1cdd9ab_1
zstd=1.4.4=ha9fde0e_3
